Hide and Edict Flag plugins
This is a modification to GNC Matt's TF2 Hide plugin. It contains changes to the sound hook and to the transmit hooks.

Included with this is an edict flag changer plugin which allows hooking of Set Transmit, particularly to hide the Heavy class.

In order to utilize this with plugins which create particle effects, you should change the edict flags of the particles within those plugins.